refactor(accessToken): move AccessTokenController and context to a new file structure
